requirements.txt
# Core dependencies
streamlit>=1.30.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0

# OCR and PDF processing
pymupdf>=1.23.0
opencv-python-headless>=4.8.0
pdf2image>=1.16.0

# Document generation
python-docx>=1.1.0

# Progress bar
tqdm>=4.66.0

# Token counting and cost estimation
tiktoken>=0.5.0

# Google Drive API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
